requirements.txt
tomli >= 2.0.0; python_version < '3.11'
requests >= 2.28.0
pytest-playwright >= 0.3.3
tomli_w >= 1.0.0